requirements.txt
# 核心依赖
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
tiktoken>=0.7.0
httpx>=0.27.0

# 图片处理（图片提取与变形修复）
Pillow>=10.0.0

# 进度与日志
tqdm>=4.66.0
rich>=13.7.0

# 脱敏处理
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354

# Streamlit GUI
streamlit>=1.35.0
streamlit-extras>=0.4.0

# 测试数据生成（开发用）
xlsxwriter>=3.2.0

